BC-5546 Refactoring
