+++ src/main/java/faang/school/analytics/config/RedisConfiguration.java
@@ -1,19 +1,30 @@
-package faang.school.analytics.config.messaging;
+package faang.school.analytics.config;
 
 import faang.school.analytics.listener.PostViewListener;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
 import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 
 @Configuration
-public class RedisPostViewConfig {
-
+public class RedisConfiguration {
+    @Value("${spring.data.redis.host}")
+    private String host;
+    @Value("${spring.data.redis.port}")
+    private int port;
     @Value("${spring.data.redis.channel.posts-view}")
-    private String channel;
+    private String postViewChannel;
+
+    @Bean
+    public JedisConnectionFactory connectionFactory() {
+        System.out.println(port);
+        RedisStandaloneConfiguration config = new RedisStandaloneConfiguration(host, port);
+        return new JedisConnectionFactory(config);
+    }
 
     @Bean
     public MessageListenerAdapter postViewAdapter(PostViewListener postViewListener) {
@@ -21,16 +32,16 @@ public class RedisPostViewConfig {
     }
 
     @Bean
-    public ChannelTopic topic() {
-        return new ChannelTopic(channel);
+    public ChannelTopic postViewTopic() {
+        return new ChannelTopic(postViewChannel);
     }
 
     @Bean
-    public RedisMessageListenerContainer redisContainer(JedisConnectionFactory connectionFactory,
-                                                        MessageListenerAdapter profileViewAdapter) {
+    public RedisMessageListenerContainer container(JedisConnectionFactory connectionFactory,
+                                                   MessageListenerAdapter profileViewAdapter) {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
         container.setConnectionFactory(connectionFactory);
-        container.addMessageListener(profileViewAdapter, topic());
+        container.addMessageListener(profileViewAdapter, postViewTopic());
         return container;
     }
 }
